+++ Sources/MapboxMaps/Foundation/MapView.swift
@@ -91,7 +91,7 @@ open class MapView: UIView {
 
     private let dependencyProvider: MapViewDependencyProviderProtocol
 
-    private var displayLinkParticipants = WeakSet<DisplayLinkParticipant>()
+    private let displayLinkParticipants = WeakSet<DisplayLinkParticipant>()
 
     /*** The preferred frames per second used for map rendering.
         NOTE: `MapView.preferredFrameRateRange` is available for iOS 15.0 and above.
@@ -300,8 +300,12 @@ open class MapView: UIView {
         // Initialize/Configure location source and location manager
         locationProducer = dependencyProvider.makeLocationProducer(
             mayRequestWhenInUseAuthorization: Bundle.main.infoDictionary?["NSLocationWhenInUseUsageDescription"] != nil)
+        let interpolatedLocationProducer = dependencyProvider.makeInterpolatedLocationProducer(
+            locationProducer: locationProducer,
+            displayLinkCoordinator: self)
         location = dependencyProvider.makeLocationManager(
             locationProducer: locationProducer,
+            interpolatedLocationProducer: interpolatedLocationProducer,
             style: mapboxMap.style)
 
         // Initialize/Configure annotations orchestrator
@@ -321,7 +325,7 @@ open class MapView: UIView {
                 anyTouchGestureRecognizer: gestures.anyTouchGestureRecognizer,
                 doubleTapGestureRecognizer: gestures.doubleTapToZoomInGestureRecognizer,
                 doubleTouchGestureRecognizer: gestures.doubleTouchToZoomOutGestureRecognizer),
-            locationProducer: locationProducer,
+            interpolatedLocationProducer: interpolatedLocationProducer,
             cameraAnimationsManager: camera,
             mapboxMap: mapboxMap)
     }