+++ Sources/MapboxMaps/Viewport/States/FollowPuck/FollowPuckViewportStateDataSource.swift
@@ -3,7 +3,7 @@ internal protocol FollowPuckViewportStateDataSourceProtocol: AnyObject {
     func observe(with handler: @escaping (CameraOptions) -> Bool) -> Cancelable
 }
 
-internal final class FollowPuckViewportStateDataSource: NSObject, FollowPuckViewportStateDataSourceProtocol {
+internal final class FollowPuckViewportStateDataSource: FollowPuckViewportStateDataSourceProtocol {
 
     internal var options: FollowPuckViewportStateOptions {
         didSet {
@@ -13,23 +13,24 @@ internal final class FollowPuckViewportStateDataSource: NSObject, FollowPuckView
 
     // MARK: - Private State
 
-    private var latestLocation: Location? {
-        didSet {
-            processUpdatedCamera()
-        }
-    }
-
+    private let interpolatedLocationProducer: InterpolatedLocationProducerProtocol
     private let observableCameraOptions: ObservableCameraOptionsProtocol
+    private let cancelables = CancelableContainer()
 
     // MARK: - Initialization
 
     internal init(options: FollowPuckViewportStateOptions,
-                  locationProducer: LocationProducerProtocol,
+                  interpolatedLocationProducer: InterpolatedLocationProducerProtocol,
                   observableCameraOptions: ObservableCameraOptionsProtocol) {
         self.options = options
+        self.interpolatedLocationProducer = interpolatedLocationProducer
         self.observableCameraOptions = observableCameraOptions
-        super.init()
-        locationProducer.add(self)
+        interpolatedLocationProducer
+            .observe { [weak self] _ in
+                self?.processUpdatedCamera()
+                return true
+            }
+            .add(to: cancelables)
     }
 
     // MARK: - Observation
@@ -41,9 +42,9 @@ internal final class FollowPuckViewportStateDataSource: NSObject, FollowPuckView
 
     // MARK: - Private Utilities
 
-    private func cameraOptions(for location: Location) -> CameraOptions {
+    private func cameraOptions(for location: InterpolatedLocation) -> CameraOptions {
         return CameraOptions(
-            center: location.location.coordinate,
+            center: location.coordinate,
             padding: options.padding,
             zoom: options.zoom,
             bearing: options.bearing?.evaluate(with: location),
@@ -51,14 +52,8 @@ internal final class FollowPuckViewportStateDataSource: NSObject, FollowPuckView
     }
 
     private func processUpdatedCamera() {
-        if let cameraOptions = latestLocation.map(cameraOptions(for:)) {
+        if let cameraOptions = interpolatedLocationProducer.location.map(cameraOptions(for:)) {
             observableCameraOptions.notify(with: cameraOptions)
         }
-    }
-}
-
-extension FollowPuckViewportStateDataSource: LocationConsumer {
-    internal func locationUpdate(newLocation: Location) {
-        latestLocation = newLocation
     }
 }