+++ Sources/MapboxMaps/Viewport/States/Helpers/ObservableCameraOptions.swift
@@ -1,5 +1,5 @@
 internal protocol ObservableCameraOptionsProtocol: AnyObject {
-    func notify(with value: CameraOptions)
+    func notify(with newValue: CameraOptions)
     func observe(with handler: @escaping (CameraOptions) -> Bool) -> Cancelable
 }
 
@@ -10,8 +10,8 @@ internal protocol ObservableCameraOptionsProtocol: AnyObject {
 internal final class ObservableCameraOptions: ObservableCameraOptionsProtocol {
     private let observableValue = ObservableValue<CameraOptions>()
 
-    internal func notify(with value: CameraOptions) {
-        observableValue.notify(with: value)
+    internal func notify(with newValue: CameraOptions) {
+        observableValue.notify(with: newValue)
     }
 
     internal func observe(with handler: @escaping (CameraOptions) -> Bool) -> Cancelable {