+++ Tests/MapboxMapsTests/Location/Puck/Puck3DTests.swift
@@ -5,20 +5,20 @@ final class Puck3DTests: XCTestCase {
 
     var configuration: Puck3DConfiguration!
     var style: MockStyle!
-    var locationProducer: MockLocationProducer!
+    var interpolatedLocationProducer: MockInterpolatedLocationProducer!
     var puck3D: Puck3D!
 
     override func setUp() {
         super.setUp()
         configuration = Puck3DConfiguration(model: Model())
         style = MockStyle()
-        locationProducer = MockLocationProducer()
+        interpolatedLocationProducer = MockInterpolatedLocationProducer()
         recreatePuck()
     }
 
     override func tearDown() {
         puck3D = nil
-        locationProducer = nil
+        interpolatedLocationProducer = nil
         style = nil
         configuration = nil
         super.tearDown()
@@ -28,7 +28,7 @@ final class Puck3DTests: XCTestCase {
         puck3D = Puck3D(
             configuration: configuration,
             style: style,
-            locationProducer: locationProducer)
+            interpolatedLocationProducer: interpolatedLocationProducer)
     }
 
     func testDefaultPropertyValues() {
@@ -37,40 +37,38 @@ final class Puck3DTests: XCTestCase {
     }
 
     func testLocationConsumerIsNotAddedAtInitialization() {
-        XCTAssertEqual(locationProducer.addStub.invocations.count, 0)
-        XCTAssertEqual(locationProducer.removeStub.invocations.count, 0)
+        XCTAssertEqual(interpolatedLocationProducer.observeStub.invocations.count, 0)
     }
 
-    func testActivatingPuckAddsLocationConsumer() {
+    func testActivatingPuckAddsLocationConsumer() throws {
         puck3D.isActive = true
 
-        XCTAssertEqual(locationProducer.addStub.invocations.count, 1)
-        XCTAssertTrue(locationProducer.addStub.parameters.first === puck3D)
-        XCTAssertEqual(locationProducer.removeStub.invocations.count, 0)
+        XCTAssertEqual(interpolatedLocationProducer.observeStub.invocations.count, 1)
+        let cancelable = try XCTUnwrap(interpolatedLocationProducer.observeStub.invocations.first?.returnValue as? MockCancelable)
+        XCTAssertEqual(cancelable.cancelStub.invocations.count, 0)
 
         // activating again should have no effect
         puck3D.isActive = true
 
-        XCTAssertEqual(locationProducer.addStub.invocations.count, 1)
-        XCTAssertEqual(locationProducer.removeStub.invocations.count, 0)
+        XCTAssertEqual(interpolatedLocationProducer.observeStub.invocations.count, 1)
+        XCTAssertEqual(cancelable.cancelStub.invocations.count, 0)
     }
 
-    func testDeactivatingPuckRemovesLocationConsumer() {
+    func testDeactivatingPuckRemovesLocationConsumer() throws {
         puck3D.isActive = true
-        locationProducer.addStub.reset()
-        locationProducer.removeStub.reset()
+        let cancelable = try XCTUnwrap(interpolatedLocationProducer.observeStub.invocations.first?.returnValue as? MockCancelable)
+        interpolatedLocationProducer.observeStub.reset()
 
         puck3D.isActive = false
 
-        XCTAssertEqual(locationProducer.addStub.invocations.count, 0)
-        XCTAssertEqual(locationProducer.removeStub.invocations.count, 1)
-        XCTAssertTrue(locationProducer.removeStub.parameters.first === puck3D)
+        XCTAssertEqual(interpolatedLocationProducer.observeStub.invocations.count, 0)
+        XCTAssertEqual(cancelable.cancelStub.invocations.count, 1)
 
         // deactivating again should have no effect
         puck3D.isActive = false
 
-        XCTAssertEqual(locationProducer.addStub.invocations.count, 0)
-        XCTAssertEqual(locationProducer.removeStub.invocations.count, 1)
+        XCTAssertEqual(interpolatedLocationProducer.observeStub.invocations.count, 0)
+        XCTAssertEqual(cancelable.cancelStub.invocations.count, 1)
     }
 
     func testSourceAndLayerAreNotAddedAtInitialization() {
@@ -80,7 +78,7 @@ final class Puck3DTests: XCTestCase {
     }
 
     func testActivatingPuckDoesNotAddSourceAndLayerIfLatestLocationIsNil() {
-        locationProducer.latestLocation = nil
+        interpolatedLocationProducer.location = nil
 
         puck3D.isActive = true
 
@@ -91,12 +89,10 @@ final class Puck3DTests: XCTestCase {
 
     func testActivatingPuckAddsSourceAndLayerIfLatestLocationIsNonNil() throws {
         let coordinate = CLLocationCoordinate2D.random()
-        locationProducer.latestLocation = Location(
-            location: CLLocation(
-                latitude: coordinate.latitude,
-                longitude: coordinate.longitude),
-            heading: nil,
-            accuracyAuthorization: .fullAccuracy)
+        var location = InterpolatedLocation.random()
+        location.coordinate = coordinate
+        location.heading = nil
+        interpolatedLocationProducer.location = location
         style.sourceExistsStub.defaultReturnValue = false
         style.layerExistsStub.defaultReturnValue = false
 
@@ -126,27 +122,17 @@ final class Puck3DTests: XCTestCase {
             .random(in: 0..<360),
             .random(in: 0..<360)]
         recreatePuck()
-        let location = CLLocation(
-            coordinate: .random(),
-            altitude: 0,
-            horizontalAccuracy: 0,
-            verticalAccuracy: 0,
-            course: .random(in: 0..<360),
-            speed: 0,
-            timestamp: Date())
-        let heading = MockHeading()
-        heading.trueHeadingStub.defaultReturnValue = .random(in: 0..<360)
-        locationProducer.latestLocation = Location(
-            location: location,
-            heading: heading,
-            accuracyAuthorization: .fullAccuracy)
+        let heading = CLLocationDirection.random(in: 0..<360)
+        var location = InterpolatedLocation.random()
+        location.heading = heading
+        interpolatedLocationProducer.location = location
         style.sourceExistsStub.defaultReturnValue = false
         puck3D.puckBearingSource = .heading
 
         puck3D.isActive = true
 
         var expectedOrientation = configuration.model.orientation!
-        expectedOrientation[2] += heading.trueHeadingStub.defaultReturnValue
+        expectedOrientation[2] += heading
         let actualSource = try XCTUnwrap(style.addSourceStub.parameters.first?.source as? ModelSource)
         XCTAssertEqual(actualSource.models?["puck-model"]?.orientation, expectedOrientation)
     }
@@ -157,27 +143,16 @@ final class Puck3DTests: XCTestCase {
             .random(in: 0..<360),
             .random(in: 0..<360)]
         recreatePuck()
-        let location = CLLocation(
-            coordinate: .random(),
-            altitude: 0,
-            horizontalAccuracy: 0,
-            verticalAccuracy: 0,
-            course: .random(in: 0..<360),
-            speed: 0,
-            timestamp: Date())
-        let heading = MockHeading()
-        heading.trueHeadingStub.defaultReturnValue = .random(in: 0..<360)
-        locationProducer.latestLocation = Location(
-            location: location,
-            heading: heading,
-            accuracyAuthorization: .fullAccuracy)
+        var location = InterpolatedLocation.random()
+        location.course = .random(in: 0..<360)
+        interpolatedLocationProducer.location = location
         style.sourceExistsStub.defaultReturnValue = false
         puck3D.puckBearingSource = .course
 
         puck3D.isActive = true
 
         var expectedOrientation = configuration.model.orientation!
-        expectedOrientation[2] += location.course
+        expectedOrientation[2] += location.course!
         let actualSource = try XCTUnwrap(style.addSourceStub.parameters.first?.source as? ModelSource)
         XCTAssertEqual(actualSource.models?["puck-model"]?.orientation, expectedOrientation)
     }
@@ -186,10 +161,7 @@ final class Puck3DTests: XCTestCase {
         configuration.modelScale = .constant(.random(withLength: 3, generator: { .random(in: 1..<10) }))
         configuration.modelRotation = .constant(.random(withLength: 3, generator: { .random(in: 0..<360) }))
         recreatePuck()
-        locationProducer.latestLocation = Location(
-            location: CLLocation(),
-            heading: nil,
-            accuracyAuthorization: .fullAccuracy)
+        interpolatedLocationProducer.location = .random()
         style.layerExistsStub.defaultReturnValue = false
 
         puck3D.isActive = true
@@ -200,20 +172,18 @@ final class Puck3DTests: XCTestCase {
     }
 
     func testUpdateExistingSource() throws {
-        let coordinate = CLLocationCoordinate2D.random()
-        locationProducer.latestLocation = Location(
-            location: CLLocation(
-                latitude: coordinate.latitude,
-                longitude: coordinate.longitude),
-            heading: nil,
-            accuracyAuthorization: .fullAccuracy)
+        var location = InterpolatedLocation.random()
+        location.heading = nil
+        interpolatedLocationProducer.location = location
         style.sourceExistsStub.defaultReturnValue = true
         style.layerExistsStub.defaultReturnValue = true
 
         puck3D.isActive = true
 
         var expectedModel = configuration.model
-        expectedModel.position = [coordinate.longitude, coordinate.latitude]
+        expectedModel.position = [
+            location.coordinate.longitude,
+            location.coordinate.latitude]
         expectedModel.orientation = [0, 0, 0]
         var expectedSource = ModelSource()
         expectedSource.models = ["puck-model": expectedModel]
@@ -230,10 +200,7 @@ final class Puck3DTests: XCTestCase {
     }
 
     func testSettingPuckBearingSourceWhenInactive() {
-        locationProducer.latestLocation = Location(
-            location: CLLocation(),
-            heading: nil,
-            accuracyAuthorization: .fullAccuracy)
+        interpolatedLocationProducer.location = .random()
         style.sourceExistsStub.defaultReturnValue = false
         style.layerExistsStub.defaultReturnValue = false
         puck3D.isActive = false
@@ -246,10 +213,7 @@ final class Puck3DTests: XCTestCase {
     }
 
     func testSettingPuckBearingSourceWhenActive() {
-        locationProducer.latestLocation = Location(
-            location: CLLocation(),
-            heading: nil,
-            accuracyAuthorization: .fullAccuracy)
+        interpolatedLocationProducer.location = .random()
         puck3D.isActive = true
         style.sourceExistsStub.defaultReturnValue = true
         style.layerExistsStub.defaultReturnValue = true
@@ -265,36 +229,19 @@ final class Puck3DTests: XCTestCase {
         XCTAssertEqual(style.addPersistentLayerWithPropertiesStub.invocations.count, 0)
     }
 
-    func testLocationUpdateWhenInactive() {
-        locationProducer.latestLocation = Location(
-            location: CLLocation(),
-            heading: nil,
-            accuracyAuthorization: .fullAccuracy)
-        style.sourceExistsStub.defaultReturnValue = false
-        style.layerExistsStub.defaultReturnValue = false
-        puck3D.isActive = false
-
-        puck3D.locationUpdate(newLocation: locationProducer.latestLocation!)
-
-        XCTAssertEqual(style.addSourceStub.invocations.count, 0)
-        XCTAssertEqual(style.addPersistentLayerStub.invocations.count, 0)
-        XCTAssertEqual(style.addPersistentLayerWithPropertiesStub.invocations.count, 0)
-    }
-
-    func testLocationUpdateWhenActive() {
-        locationProducer.latestLocation = Location(
-            location: CLLocation(),
-            heading: nil,
-            accuracyAuthorization: .fullAccuracy)
+    func testLocationUpdateWhenActive() throws {
+        interpolatedLocationProducer.location = .random()
         puck3D.isActive = true
         style.sourceExistsStub.defaultReturnValue = true
         style.layerExistsStub.defaultReturnValue = true
         style.addSourceStub.reset()
         style.setSourcePropertiesStub.reset()
         style.addPersistentLayerStub.reset()
+        let handler = try XCTUnwrap(interpolatedLocationProducer.observeStub.invocations.first?.parameters)
 
-        puck3D.locationUpdate(newLocation: locationProducer.latestLocation!)
+        let wantsMoreUpdates = handler(interpolatedLocationProducer.location!)
 
+        XCTAssertTrue(wantsMoreUpdates)
         XCTAssertEqual(style.addSourceStub.invocations.count, 0)
         XCTAssertEqual(style.setSourcePropertiesStub.invocations.count, 1)
         XCTAssertEqual(style.addPersistentLayerStub.invocations.count, 0)